requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
PyAudio==0.2.9
python-Levenshtein==0.12.0
requests==2.13.0
Unidecode==0.4.20
googletrans
decorator
